requirements.txt
# 基础环境依赖
python-dotenv==1.0.0
python-telegram-bot==20.3
nest-asyncio==1.5.6
requests==2.31.0
volcengine-python-sdk[ark]==1.0.128
groq==0.19.0